+++ scheduler/models/base_model.py
@@ -1,5 +1,4 @@
 # Path: scheduler/models/base_model.py
 
 class BaseModel:
-    def __init__(self, db):
-        self.db = db
+    pass